Add dorky C++ code to possibly parse .root files
The goal here is to easily analyze the performance_ckf.root
file, and the answer is obviously uproot, but I also want
to know how to work with .root files in C++ so I made this
dorky thing to start with. Either I parse the .root file
in C++, Root Script, TBrowser, or Python. Whichever is
easiest I suppose.